requirements.txt
torch==2.5.1
torchvision==0.20.1
numpy==1.26.4
einops
pillow
imageio-ffmpeg
scikit-image
opencv-python
plyfile
huggingface_hub
safetensors
matplotlib
tqdm
pandas
rootutils
hydra-core==1.3.2
hydra-colorlog
open3d  # for point map evaluation
evo==1.31.0  # for camera distance evaluation
pytubefix  # for re10k downloader
h5py  # for nyu-v2 preprocessing